ginkgo: bring back smart charging
[@nckmml]
As I have noticed maintainers are removing smart charging on several unaffected devices. I had a conversation with kubersharma, who announced the removal for vayu and made the xda post many people are referencing, he assured me that his post was mostly for vayu devices, he is only aware of the bug occurring on vayu and surya at all. There is no reason to assume that it would be generally unsafe for all other devices. Maintainers are supposed to decide on device base if something like that occured before and if they should remove it, otherwise its a really useful feature that can protect many peoples batteries from early degrading.

This reverts commit 8e83ca7f3f7b5186cc7d13cb6f6cb626449741b3.